front/requirements.txt
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.4
dnspython==2.2.0
email-validator==1.1.3
filelock==3.4.2
Flask==2.0.2
Flask-Compress==1.10.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.29.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
imageio==2.14.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
networkx==2.6.3
numpy==1.22.1
outcome==1.1.0
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
pipenv==2022.1.8
platformdirs==2.4.1
plotly==5.5.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
requests==2.27.1
scikit-image==0.19.1
scipy==1.7.3
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
tenacity==8.0.1
tifffile==2021.11.2
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
Werkzeug==2.0.2
wsproto==1.0.0
WTForms==3.0.1
